adding more shared UI
