+++ inspector/src/components/actionTabs/tabs/propertyGrids/customPropertyGridComponent.tsx
@@ -1,7 +1,7 @@
 import * as React from "react";
 import { Observable } from "babylonjs/Misc/observable";
 import { PropertyChangedEvent } from "../../../propertyChangedEvent";
-import { Color3LineComponent } from "../../lines/color3LineComponent";
+import { Color3LineComponent } from "../../../../sharedUiComponents/lines/color3LineComponent";
 import { GlobalState } from '../../../globalState';
 import { IInspectable, InspectableType } from 'babylonjs/Misc/iInspectable';
 import { CheckBoxLineComponent } from '../../../../sharedUiComponents/lines/checkBoxLineComponent';
