+++ inspector/src/components/actionTabs/tabs/propertyGrids/gui/imagePropertyGridComponent.tsx
@@ -7,7 +7,7 @@ import { LockObject } from "../lockObject";
 import { Image } from "babylonjs-gui/2D/controls/image";
 import { FloatLineComponent } from "../../../lines/floatLineComponent";
 import { CheckBoxLineComponent } from "../../../../../sharedUiComponents/lines/checkBoxLineComponent";
-import { OptionsLineComponent } from "../../../lines/optionsLineComponent";
+import { OptionsLineComponent } from "../../../../../sharedUiComponents/lines/optionsLineComponent";
 import { GlobalState } from '../../../../globalState';
 
 interface IImagePropertyGridComponentProps {
