+++ inspector/src/components/actionTabs/tabs/propertyGrids/materials/nodeMaterialPropertyGridComponent.tsx
@@ -2,7 +2,6 @@ import * as React from "react";
 
 import { Observable } from "babylonjs/Misc/observable";
 import { NodeMaterial } from "babylonjs/Materials/Node/nodeMaterial";
-
 import { PropertyChangedEvent } from "../../../../propertyChangedEvent";
 import { LineContainerComponent } from "../../../lineContainerComponent";
 import { CommonMaterialPropertyGridComponent } from "./commonMaterialPropertyGridComponent";
@@ -11,16 +10,16 @@ import { GlobalState } from '../../../../globalState';
 import { ButtonLineComponent } from '../../../../../sharedUiComponents/lines/buttonLineComponent';
 import { CheckBoxLineComponent } from '../../../../../sharedUiComponents/lines/checkBoxLineComponent';
 import { FloatLineComponent } from '../../../lines/floatLineComponent';
-import { Color3LineComponent } from '../../../lines/color3LineComponent';
+import { Color3LineComponent } from '../../../../../sharedUiComponents/lines/color3LineComponent';
 import { Vector3LineComponent } from '../../../lines/vector3LineComponent';
-import { Vector4LineComponent } from '../../../lines/vector4LineComponent';
-import { Vector2LineComponent } from '../../../lines/vector2LineComponent';
+import { Vector4LineComponent } from '../../../../../sharedUiComponents/lines/vector4LineComponent';
+import { Vector2LineComponent } from '../../../../../sharedUiComponents/lines/vector2LineComponent';
 import { TextureLinkLineComponent } from '../../../lines/textureLinkLineComponent';
 import { SliderLineComponent } from '../../../lines/sliderLineComponent';
 import { NodeMaterialBlockConnectionPointTypes } from 'babylonjs/Materials/Node/Enums/nodeMaterialBlockConnectionPointTypes';
 import { InputBlock } from 'babylonjs/Materials/Node/Blocks/Input/inputBlock';
-import { Color4LineComponent } from '../../../lines/color4LineComponent';
 import { GradientPropertyTabComponent } from '../../gradientNodePropertyComponent';
+import { Color4LineComponent } from "../../../../../sharedUiComponents/lines/color4LineComponent";
 
 interface INodeMaterialPropertyGridComponentProps {
     globalState: GlobalState;
