+++ inspector/src/components/actionTabs/tabs/propertyGrids/particleSystems/particleSystemPropertyGridComponent.tsx
@@ -1,7 +1,6 @@
 import * as React from "react";
 
 import { Observable } from "babylonjs/Misc/observable";
-
 import { PropertyChangedEvent } from "../../../../propertyChangedEvent";
 import { LineContainerComponent } from "../../../lineContainerComponent";
 import { TextLineComponent } from "../../../../../sharedUiComponents/lines/textLineComponent";
@@ -12,9 +11,8 @@ import { IParticleSystem } from 'babylonjs/Particles/IParticleSystem';
 import { FloatLineComponent } from '../../../lines/floatLineComponent';
 import { ButtonLineComponent } from '../../../../../sharedUiComponents/lines/buttonLineComponent';
 import { TextureLinkLineComponent } from '../../../lines/textureLinkLineComponent';
-import { OptionsLineComponent } from '../../../lines/optionsLineComponent';
+import { OptionsLineComponent } from '../../../../../sharedUiComponents/lines/optionsLineComponent';
 import { ParticleSystem } from 'babylonjs/Particles/particleSystem';
-import { Color4LineComponent } from '../../../lines/color4LineComponent';
 import { Vector3LineComponent } from '../../../lines/vector3LineComponent';
 import { CheckBoxLineComponent } from '../../../../../sharedUiComponents/lines/checkBoxLineComponent';
 import { SliderLineComponent } from '../../../lines/sliderLineComponent';
@@ -41,6 +39,7 @@ import { Tools } from 'babylonjs/Misc/tools';
 import { FileButtonLineComponent } from '../../../../../sharedUiComponents/lines/fileButtonLineComponent';
 import { TextInputLineComponent } from '../../../lines/textInputLineComponent';
 import { ParticleHelper } from 'babylonjs/Particles/particleHelper';
+import { Color4LineComponent } from "../../../../../sharedUiComponents/lines/color4LineComponent";
 
 interface IParticleSystemPropertyGridComponentProps {
     globalState: GlobalState;
