+++ inspector/src/components/actionTabs/tabs/propertyGrids/postProcesses/defaultRenderingPipelinePropertyGridComponent.tsx
@@ -9,12 +9,12 @@ import { CommonRenderingPipelinePropertyGridComponent } from './commonRenderingP
 import { SliderLineComponent } from '../../../lines/sliderLineComponent';
 import { LineContainerComponent } from '../../../lineContainerComponent';
 import { CheckBoxLineComponent } from '../../../../../sharedUiComponents/lines/checkBoxLineComponent';
-import { Vector2LineComponent } from '../../../lines/vector2LineComponent';
-import { OptionsLineComponent } from '../../../lines/optionsLineComponent';
+import { OptionsLineComponent } from '../../../../../sharedUiComponents/lines/optionsLineComponent';
 import { ImageProcessingConfiguration } from 'babylonjs/Materials/imageProcessingConfiguration';
-import { Color3LineComponent } from '../../../lines/color3LineComponent';
+import { Color3LineComponent } from '../../../../../sharedUiComponents/lines/color3LineComponent';
 import { GlobalState } from '../../../../globalState';
 import { ButtonLineComponent } from '../../../../../sharedUiComponents/lines/buttonLineComponent';
+import { Vector2LineComponent } from "../../../../../sharedUiComponents/lines/vector2LineComponent";
 
 interface IDefaultRenderingPipelinePropertyGridComponentProps {
     globalState: GlobalState;
