+++ inspector/src/components/actionTabs/tabs/tools/gltfComponent.tsx
@@ -4,7 +4,7 @@ import { LineContainerComponent } from "../../lineContainerComponent";
 import { CheckBoxLineComponent } from "../../../../sharedUiComponents/lines/checkBoxLineComponent";
 import { GlobalState } from "../../../globalState";
 import { FloatLineComponent } from "../../lines/floatLineComponent";
-import { OptionsLineComponent } from "../../lines/optionsLineComponent";
+import { OptionsLineComponent } from "../../../../sharedUiComponents/lines/optionsLineComponent";
 import { MessageLineComponent } from "../../../../sharedUiComponents/lines/messageLineComponent";
 import { faCheck, faTimesCircle } from "@fortawesome/free-solid-svg-icons";
 import { TextLineComponent } from "../../../../sharedUiComponents/lines/textLineComponent";
